+++ lib/searchkick/index.rb
@@ -26,7 +26,7 @@ module Searchkick
       client.index(
         index: name,
         type: document_type(record),
-        id: record.id,
+        id: search_id(record),
         body: search_data(record)
       )
     end
@@ -35,16 +35,16 @@ module Searchkick
       client.delete(
         index: name,
         type: document_type(record),
-        id: record.id
+        id: search_id(record)
       )
     end
 
     def import(records)
-      if records.any?
+      records.group_by{|r| document_type(r) }.each do |type, batch|
         client.bulk(
           index: name,
-          type: document_type(records.first),
-          body: records.map{|r| data = search_data(r); {index: {_id: data["_id"] || data["id"] || r.id, data: data}} }
+          type: type,
+          body: batch.map{|r| {index: {_id: search_id(r), data: search_data(r)}} }
         )
       end
     end
@@ -58,7 +58,11 @@ module Searchkick
     end
 
     def klass_document_type(klass)
-      klass.model_name.to_s.underscore
+      if klass.respond_to?(:document_type)
+        klass.document_type
+      else
+        klass.model_name.to_s.underscore
+      end
     end
 
     protected
@@ -71,18 +75,17 @@ module Searchkick
       klass_document_type(record.class)
     end
 
+    def search_id(record)
+      record.id.is_a?(Numeric) ? record.id : record.id.to_s
+    end
+
     def search_data(record)
       source = record.search_data
+      options = record.class.searchkick_options
 
       # stringify fields
-      source = source.inject({}){|memo,(k,v)| memo[k.to_s] = v; memo}
-
-      # Mongoid 4 hack
-      if defined?(BSON::ObjectId) and source["_id"].is_a?(BSON::ObjectId)
-        source["_id"] = source["_id"].to_s
-      end
-
-      options = record.class.searchkick_options
+      # remove _id since search_id is used instead
+      source = source.inject({}){|memo,(k,v)| memo[k.to_s] = v; memo}.except("_id")
 
       # conversions
       conversions_field = options[:conversions]
@@ -107,8 +110,6 @@ module Searchkick
       end
 
       cast_big_decimal(source)
-
-      # p search_data
 
       source.as_json
     end